+++ src/main/kotlin/pullrequestfactory/domain/branches/Branch.kt
@@ -1,5 +1,8 @@
 package pullrequestfactory.domain.branches
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class Branch(val name: String) {
 
     fun iterationNr(): Int = parts().dropLast(1).last().toInt()