+++ src/main/kotlin/pullrequestfactory/domain/branches/Branches.kt
@@ -16,8 +16,8 @@ class Branches(
             val prs = branches.mapIndexed { idx, branch ->
                 PullRequest(
                     title = getBranchTitles()[idx],
-                    _base = getBaseBranches()[idx].copy(),
-                    _head = branch.copy()
+                    base = getBaseBranches()[idx].copy(),
+                    head = branch.copy()
                 )
             }
             return prMarker.markTitlesOf(prs)
@@ -69,8 +69,8 @@ class Branches(
     private fun getBranchTitles(): List<String> = branches.mapIndexed { idx, br ->
         val (firstName, lastName, _, iterationNr, pairingPartner) = br.parts()
         val titleParts = listOf(
-            firstName.capitalize(),
-            lastName.capitalize(),
+            firstName.replaceFirstChar { it.titlecase() },
+            lastName.replaceFirstChar { it.titlecase() },
             "Iteration",
             iterationNr,
             "/",