+++ src/main/kotlin/pullrequestfactory/domain/pullrequests/GetPullRequest.kt
@@ -1,3 +1,6 @@
 package pullrequestfactory.domain.pullrequests
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class GetPullRequest(val number: Int, val title: String)