+++ src/main/kotlin/pullrequestfactory/io/clikt/OpenCommand.kt
@@ -40,7 +40,7 @@ class OpenCommand(
         val pps = listOf(pp1, pp2, pp3, pp4, pp5, pp6, pp7)
         OpenPullRequestsProgram(
             args.ui,
-            GithubAPIClient(httpClient, args.baseUrl, args.repoUrl, args.ui),
+            GithubAPIClient(httpClient, args.baseUrl, args.repoUrl, args.ui, args.jsonSerizalizer),
             httpClient,
             isLastFinished,
             candidate,